backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
beautifulsoup4==4.14.3
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.1.8
distro==1.9.0
exceptiongroup==1.3.1
fastapi==0.128.0
filelock==3.19.1
fsspec==2025.10.0
groq==1.0.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
packaging==26.0
pillow==11.3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
safetensors==0.7.0
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==5.1.2
sniffio==1.3.1
soupsieve==2.8.3
starlette==0.49.3
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.8.0
tqdm==4.67.1
transformers==4.57.6
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.39.0
